move sdk join tests
